Release/1.2.0

+++ src/PriceCalculator.Api/Bll/Models/GoodModel.cs
@@ -1,8 +1,8 @@
 ﻿namespace PriceCalculator.Api.Bll.Models;
 
 public record GoodModel(
-    int Length,
-    int Width,
-    int Height,
-    double Weight = 0
+    decimal Length,
+    decimal Width,
+    decimal Height,
+    decimal Weight = 0
     );
